Move benchmarks from testnet config to develop
