+++ Cargo.toml
@@ -141,7 +141,7 @@ default_identity = [
     "polymesh-common-utilities/default_identity"
 ]
 runtime-benchmarks = [
-    "polymesh-runtime-testnet/runtime-benchmarks"
+    "polymesh-runtime-develop/runtime-benchmarks"
 ]
 std = [
     "serde/std",
