+++ pallets/runtime/develop/Cargo.toml
@@ -89,6 +89,11 @@ pallet-contracts-rpc-runtime-api = { git = "https://github.com/paritytech/substr
 pallet-staking-rpc-runtime-api = { path = "../../staking/rpc/runtime-api", default-features = false }
 node-rpc-runtime-api = { path = "../../../rpc/runtime-api", default-features = false }
 
+# Used for runtime benchmarking
+frame-benchmarking = { default-features = false, git = "https://github.com/paritytech/substrate", tag = "v2.0.0", optional = true }
+frame-system-benchmarking = { default-features = false, git = "https://github.com/paritytech/substrate", tag = "v2.0.0", optional = true }
+hex-literal = { version = "0.3.1", optional = true }
+
 [build-dependencies]
 wasm-builder-runner = { package = "substrate-wasm-builder-runner", git = "https://github.com/paritytech/substrate", default-features = false, tag = "v2.0.0"}
 
@@ -101,7 +106,9 @@ no_std = [
 ]
 std = [
     "codec/std",
+    "frame-benchmarking/std",
     "frame-support/std",
+    "frame-system-benchmarking/std",
     "frame-system-rpc-runtime-api/std",
     "frame-system/std",
     "node-rpc-runtime-api/std",
@@ -167,4 +174,22 @@ std = [
     "sp-transaction-pool/std",
     "sp-version/std",
     "polymesh-contracts/std",
+]
+runtime-benchmarks = [
+    "frame-benchmarking",
+    "frame-support/runtime-benchmarks",
+    "frame-system-benchmarking",
+    "frame-system/runtime-benchmarks",
+    "hex-literal",
+    "pallet-asset/runtime-benchmarks",
+    "pallet-balances/runtime-benchmarks",
+    "pallet-committee/runtime-benchmarks",
+    "pallet-identity/runtime-benchmarks",
+    "pallet-im-online/runtime-benchmarks",
+    "pallet-portfolio/runtime-benchmarks",
+    "pallet-staking/runtime-benchmarks",
+    "pallet-timestamp/runtime-benchmarks",
+    "polymesh-primitives/runtime-benchmarks",
+    #"polymesh-runtime-common/runtime-benchmarks",
+    "sp-runtime/runtime-benchmarks",
 ]
